refactor(ts): migrate server
