+++ src/server.ts
@@ -1,5 +1,5 @@
 import URI from 'urijs';
-import { xdr, StrKey, Asset } from 'stellar-base';
+import { xdr, StrKey, Asset, Transaction, AssetType } from 'stellar-base';
 import BigNumber from 'bignumber.js';
 
 import { BadResponseError } from './errors';
@@ -23,12 +23,14 @@ import { EffectCallBuilder } from './effect_call_builder';
 import { FriendbotBuilder } from './friendbot_builder';
 import { AssetsCallBuilder } from './assets_call_builder';
 import { TradeAggregationCallBuilder } from './trade_aggregation_call_builder';
+import { Horizon } from './horizon_api';
+import { Omit } from 'lodash';
 
 export const SUBMIT_TRANSACTION_TIMEOUT = 60 * 1000;
 
 const STROOPS_IN_LUMEN = 10000000;
 
-function _getAmountInLumens(amt) {
+function _getAmountInLumens(amt: BigNumber) {
   return new BigNumber(amt).div(STROOPS_IN_LUMEN).toString();
 }
 
@@ -41,13 +43,19 @@ function _getAmountInLumens(amt) {
  * @param {boolean} [opts.allowHttp] - Allow connecting to http servers, default: `false`. This must be set to false in production deployments! You can also use {@link Config} class to set this globally.
  */
 export class Server {
-  constructor(serverURL, opts = {}) {
+  /**
+   * serverURL Horizon Server URL (ex. `https://horizon-testnet.stellar.org`).
+   *
+   * TODO: Solve `URI(this.serverURL as any)`.
+   */
+  public readonly serverURL: uri.URI
+
+  constructor(serverURL: string, opts: Server.Options = {}) {
     this.serverURL = URI(serverURL);
 
-    let allowHttp = Config.isAllowHttp();
-    if (typeof opts.allowHttp !== 'undefined') {
-      allowHttp = opts.allowHttp;
-    }
+    const allowHttp = typeof opts.allowHttp === 'undefined'
+      ? Config.isAllowHttp()
+      : opts.allowHttp;
 
     if (this.serverURL.protocol() !== 'https' && !allowHttp) {
       throw new Error('Cannot connect to insecure horizon server');
@@ -84,30 +92,29 @@ export class Server {
    * @returns {Promise<Timebounds>} Promise that resolves a `timebounds` object
    * (with the shape `{ minTime: 0, maxTime: N }`) that you can set the `timebounds` option to.
    */
-  fetchTimebounds(seconds, _isRetry = false) {
+  public async fetchTimebounds(seconds: number, _isRetry: boolean = false): Promise<Server.Timebounds> {
     // HorizonAxiosClient instead of this.ledgers so we can get at them headers
     const currentTime = getCurrentServerTime(this.serverURL.hostname());
 
     if (currentTime) {
-      return Promise.resolve({
+      return {
         minTime: 0,
         maxTime: currentTime + seconds
-      });
+      };
     }
 
     // if this is a retry, then the retry has failed, so use local time
     if (_isRetry) {
-      return Promise.resolve({
+      return {
         minTime: 0,
         maxTime: Math.floor(new Date().getTime() / 1000) + seconds
-      });
+      };
     }
 
     // otherwise, retry (by calling the root endpoint)
     // toString automatically adds the trailing slash
-    return HorizonAxiosClient.get(URI(this.serverURL).toString()).then(() =>
-      this.fetchTimebounds(seconds, true)
-    );
+    await HorizonAxiosClient.get(URI(this.serverURL as any).toString());
+    return await this.fetchTimebounds(seconds, true);
   }
 
   /**
@@ -116,17 +123,15 @@ export class Server {
    * that happens!
    * @returns {Promise<number>} Promise that resolves to the base fee.
    */
-  fetchBaseFee() {
-    return this.ledgers()
+  public async fetchBaseFee(): Promise<number> {
+    const response = await this.ledgers()
       .order('desc')
       .limit(1)
-      .call()
-      .then((response) => {
-        if (response && response.records[0]) {
-          return response.records[0].base_fee_in_stroops || 100;
-        }
-        return 100;
-      });
+      .call();
+    if (response && response.records && response.records[0]) {
+      return response.records[0].base_fee_in_stroops || 100;
+    }
+    return 100;
   }
 
   /**
@@ -134,8 +139,8 @@ export class Server {
    * @see [Operation Fee Stats](https://www.stellar.org/developers/horizon/reference/endpoints/fee-stats.html)
    * @returns {Promise} Promise that resolves to the fee stats returned by Horizon.
    */
-  operationFeeStats() {
-    const cb = new CallBuilder(URI(this.serverURL));
+  public async operationFeeStats(): Promise<any> {
+    const cb = new CallBuilder(this.serverURL);
     cb.filter.push(['operation_fee_stats']);
     return cb.call();
   }
@@ -234,7 +239,7 @@ export class Server {
    * @param {Transaction} transaction - The transaction to submit.
    * @returns {Promise} Promise that resolves or rejects with response from horizon.
    */
-  submitTransaction(transaction) {
+  public async submitTransaction(transaction: Transaction): Promise<Server.TransactionRecord> {
     const tx = encodeURIComponent(
       transaction
         .toEnvelope()
@@ -243,7 +248,7 @@ export class Server {
     );
 
     return HorizonAxiosClient.post(
-      URI(this.serverURL)
+      URI(this.serverURL.toString())  // TODO: Is toString() neccessary?)
         .segment('transactions')
         .toString(),
       `tx=${tx}`,
@@ -254,19 +259,22 @@ export class Server {
           return response.data;
         }
 
-        const responseXDR = xdr.TransactionResult.fromXDR(
+        // TODO: fix stellar-base types.
+        const responseXDR: xdr.TransactionResult = (xdr.TransactionResult.fromXDR as any)(
           response.data.result_xdr,
           'base64'
         );
 
-        const results = responseXDR.result().value();
+        // TODO: fix stellar-base types.
+        const results = (responseXDR as any).result().value();
 
         let offerResults;
         let hasManageOffer;
 
         if (results.length) {
           offerResults = results
-            .map((result, i) => {
+            // TODO: fix stellar-base types.
+            .map((result: any, i: number) => {
               if (
                 result.value().switch().name !== 'manageBuyOffer' &&
                 result.value().switch().name !== 'manageSellOffer'
@@ -286,7 +294,8 @@ export class Server {
 
               const offersClaimed = offerSuccess
                 .offersClaimed()
-                .map((offerClaimed) => {
+                // TODO: fix stellar-base types.
+                .map((offerClaimed: any) => {
                   const claimedOfferAmountBought = new BigNumber(
                     // amountBought is a js-xdr hyper
                     offerClaimed.amountBought().toString()
@@ -391,7 +400,8 @@ export class Server {
                   !offersClaimed.length && effect === 'manageOfferDeleted'
               };
             })
-            .filter((result) => !!result);
+            // TODO: fix stellar-base types.
+            .filter((result: any) => !!result);
         }
 
         return Object.assign({}, response.data, {
@@ -416,22 +426,22 @@ export class Server {
   /**
    * @returns {AccountCallBuilder} New {@link AccountCallBuilder} object configured by a current Horizon server configuration.
    */
-  accounts() {
-    return new AccountCallBuilder(URI(this.serverURL));
+  public accounts(): AccountCallBuilder {
+    return new AccountCallBuilder(URI(this.serverURL as any));
   }
 
   /**
    * @returns {LedgerCallBuilder} New {@link LedgerCallBuilder} object configured by a current Horizon server configuration.
    */
-  ledgers() {
-    return new LedgerCallBuilder(URI(this.serverURL));
+  public ledgers(): LedgerCallBuilder {
+    return new LedgerCallBuilder(URI(this.serverURL as any));
   }
 
   /**
    * @returns {TransactionCallBuilder} New {@link TransactionCallBuilder} object configured by a current Horizon server configuration.
    */
-  transactions() {
-    return new TransactionCallBuilder(URI(this.serverURL));
+  public transactions(): TransactionCallBuilder {
+    return new TransactionCallBuilder(URI(this.serverURL as any));
   }
 
   /**
@@ -447,9 +457,9 @@ export class Server {
    * @param {...string} resourceParams Parameters for selected resource
    * @returns {OfferCallBuilder} New {@link OfferCallBuilder} object
    */
-  offers(resource, ...resourceParams) {
+  public offers(resource: string, ...resourceParams: string[]): OfferCallBuilder {
     return new OfferCallBuilder(
-      URI(this.serverURL),
+      this.serverURL,
       resource,
       ...resourceParams
     );
@@ -460,23 +470,23 @@ export class Server {
    * @param {Asset} buying Asset being bought
    * @returns {OrderbookCallBuilder} New {@link OrderbookCallBuilder} object configured by a current Horizon server configuration.
    */
-  orderbook(selling, buying) {
-    return new OrderbookCallBuilder(URI(this.serverURL), selling, buying);
+  public orderbook(selling: Asset, buying: Asset): OrderbookCallBuilder {
+    return new OrderbookCallBuilder(URI(this.serverURL as any), selling, buying);
   }
 
   /**
    * Returns
    * @returns {TradesCallBuilder} New {@link TradesCallBuilder} object configured by a current Horizon server configuration.
    */
-  trades() {
-    return new TradesCallBuilder(URI(this.serverURL));
+  public trades(): TradesCallBuilder {
+    return new TradesCallBuilder(URI(this.serverURL as any));
   }
 
   /**
    * @returns {OperationCallBuilder} New {@link OperationCallBuilder} object configured by a current Horizon server configuration.
    */
-  operations() {
-    return new OperationCallBuilder(URI(this.serverURL));
+  public operations(): OperationCallBuilder {
+    return new OperationCallBuilder(URI(this.serverURL as any));
   }
 
   /**
@@ -500,9 +510,9 @@ export class Server {
    * @param {string} destinationAmount The amount, denominated in the destination asset, that any returned path should be able to satisfy.
    * @returns {PathCallBuilder} New {@link PathCallBuilder} object configured with the current Horizon server configuration.
    */
-  paths(source, destination, destinationAsset, destinationAmount) {
+  public paths(source: string, destination: string, destinationAsset: Asset, destinationAmount: string): PathCallBuilder {
     return new PathCallBuilder(
-      URI(this.serverURL),
+      this.serverURL,
       source,
       destination,
       destinationAsset,
@@ -514,16 +524,16 @@ export class Server {
    * @returns {PaymentCallBuilder} New {@link PaymentCallBuilder} instance configured with the current
    * Horizon server configuration.
    */
-  payments() {
-    return new PaymentCallBuilder(URI(this.serverURL));
+  public payments(): PaymentCallBuilder {
+    return new PaymentCallBuilder(URI(this.serverURL as any) as any);
   }
 
   /**
    * @returns {EffectCallBuilder} New {@link EffectCallBuilder} instance configured with the current
    * Horizon server configuration
    */
-  effects() {
-    return new EffectCallBuilder(URI(this.serverURL));
+  public effects(): EffectCallBuilder {
+    return new EffectCallBuilder(URI(this.serverURL as any) as any);
   }
 
   /**
@@ -532,8 +542,8 @@ export class Server {
    * Horizon server configuration
    * @private
    */
-  friendbot(address) {
-    return new FriendbotBuilder(URI(this.serverURL), address);
+  public friendbot(address: string): FriendbotBuilder {
+    return new FriendbotBuilder(URI(this.serverURL as any), address);  // TODO: Is toString() neccessary?
   }
 
   /**
@@ -541,8 +551,8 @@ export class Server {
    * Horizon server configuration.
    * @returns {AssetsCallBuilder} New AssetsCallBuilder instance
    */
-  assets() {
-    return new AssetsCallBuilder(URI(this.serverURL));
+  public assets(): AssetsCallBuilder {
+    return new AssetsCallBuilder(URI(this.serverURL as any));
   }
 
   /**
@@ -550,11 +560,11 @@ export class Server {
    * @param {string} accountId - The account to load.
    * @returns {Promise} Returns a promise to the {@link AccountResponse} object with populated sequence number.
    */
-  loadAccount(accountId) {
-    return this.accounts()
+  public async loadAccount(accountId: string): Promise<AccountResponse> {
+    const res = await this.accounts()
       .accountId(accountId)
-      .call()
-      .then((res) => new AccountResponse(res));
+      .call();
+    return new AccountResponse(res);
   }
 
   /**
@@ -568,9 +578,9 @@ export class Server {
    * Returns new {@link TradeAggregationCallBuilder} object configured with the current Horizon server configuration.
    * @returns {TradeAggregationCallBuilder} New TradeAggregationCallBuilder instance
    */
-  tradeAggregation(base, counter, start_time, end_time, resolution, offset) {
+  tradeAggregation(base: Asset, counter: Asset, start_time: number, end_time: number, resolution: number, offset: number): TradeAggregationCallBuilder {
     return new TradeAggregationCallBuilder(
-      URI(this.serverURL),
+      this.serverURL,
       base,
       counter,
       start_time,
@@ -579,4 +589,268 @@ export class Server {
       offset
     );
   }
+}
+
+export namespace Server {
+  export interface Options {
+    allowHttp?: boolean;
+  }
+
+  export interface Timebounds {
+    minTime: number;
+    maxTime: number;
+  }
+
+
+  /* Due to a bug with the recursive function requests */
+  interface CollectionRecord<
+    T extends Horizon.BaseResponse = Horizon.BaseResponse
+  > {
+    _links: {
+      next: Horizon.ResponseLink;
+      prev: Horizon.ResponseLink;
+      self: Horizon.ResponseLink;
+    };
+    _embedded: {
+      records: T[];
+    };
+  }
+
+  interface CallFunctionTemplateOptions {
+    cursor?: string | number;
+    limit?: number;
+    order?: 'asc' | 'desc';
+  }
+
+  type CallFunction<
+    T extends Horizon.BaseResponse = Horizon.BaseResponse
+  > = () => Promise<T>;
+  type CallCollectionFunction<
+    T extends Horizon.BaseResponse = Horizon.BaseResponse
+  > = (options?: CallFunctionTemplateOptions) => Promise<CollectionRecord<T>>;
+
+  interface AccountRecord extends Horizon.BaseResponse {
+    id: string;
+    paging_token: string;
+    account_id: string;
+    sequence: number;
+    subentry_count: number;
+    inflation_destination?: string;
+    last_modified_ledger: number;
+    thresholds: Horizon.AccountThresholds;
+    flags: Horizon.Flags;
+    balances: Horizon.BalanceLine[];
+    signers: Array<{
+      key: string;
+      weight: number;
+      type: string;
+    }>;
+    data: (options: { value: string }) => Promise<{ value: string }>;
+    data_attr: {
+      [key: string]: string;
+    };
+    effects: CallCollectionFunction<EffectRecord>;
+    offers: CallCollectionFunction<OfferRecord>;
+    operations: CallCollectionFunction<OperationRecord>;
+    payments: CallCollectionFunction<PaymentOperationRecord>;
+    trades: CallCollectionFunction<TradeRecord>;
+  }
+
+  interface EffectRecord extends Horizon.BaseResponse {
+    account: string;
+    paging_token: string;
+    starting_balance: string;
+    type_i: string;
+    type: string;
+    amount?: any;
+
+    operation?: CallFunction<OperationRecord>;
+    precedes?: CallFunction<EffectRecord>;
+    succeeds?: CallFunction<EffectRecord>;
+  }
+
+  interface LedgerRecord extends Horizon.BaseResponse {
+    id: string;
+    paging_token: string;
+    hash: string;
+    prev_hash: string;
+    sequence: number;
+    transaction_count: number;
+    operation_count: number;
+    closed_at: string;
+    total_coins: string;
+    fee_pool: string;
+    base_fee: number;
+    base_reserve: string;
+    max_tx_set_size: number;
+    protocol_version: number;
+    header_xdr: string;
+    base_fee_in_stroops: number;
+    base_reserve_in_stroops: number;
+
+    effects: CallCollectionFunction<EffectRecord>;
+    operations: CallCollectionFunction<OperationRecord>;
+    self: CallFunction<LedgerRecord>;
+    transactions: CallCollectionFunction<TransactionRecord>;
+  }
+
+  interface OfferAsset {
+    asset_type: AssetType;
+    asset_code?: string;
+    asset_issuer?: string;
+  }
+
+  interface OfferRecord extends Horizon.BaseResponse {
+    id: string;
+    paging_token: string;
+    seller: string;
+    selling: OfferAsset;
+    buying: OfferAsset;
+    amount: string;
+    price_r: Horizon.PriceRShorthand;
+    price: string;
+    last_modified_ledger: number;
+    last_modified_time: string;
+  }
+
+  import OperationResponseType = Horizon.OperationResponseType;
+  import OperationResponseTypeI = Horizon.OperationResponseTypeI;
+  interface BaseOperationRecord<
+    T extends OperationResponseType = OperationResponseType,
+    TI extends OperationResponseTypeI = OperationResponseTypeI
+  > extends Horizon.BaseOperationResponse<T, TI> {
+    self: CallFunction<OperationRecord>;
+    succeeds: CallFunction<OperationRecord>;
+    precedes: CallFunction<OperationRecord>;
+    effects: CallCollectionFunction<EffectRecord>;
+    transaction: CallFunction<TransactionRecord>;
+  }
+
+  interface CreateAccountOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.createAccount,
+        OperationResponseTypeI.createAccount
+      >,
+      Horizon.CreateAccountOperationResponse {}
+  interface PaymentOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.payment,
+        OperationResponseTypeI.payment
+      >,
+      Horizon.PaymentOperationResponse {
+    sender: CallFunction<AccountRecord>;
+    receiver: CallFunction<AccountRecord>;
+  }
+  interface PathPaymentOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.pathPayment,
+        OperationResponseTypeI.pathPayment
+      >,
+      Horizon.PathPaymentOperationResponse {}
+  interface ManageOfferOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.manageOffer,
+        OperationResponseTypeI.manageOffer
+      >,
+      Horizon.ManageOfferOperationResponse {}
+  interface PassiveOfferOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.createPassiveOffer,
+        OperationResponseTypeI.createPassiveOffer
+      >,
+      Horizon.PassiveOfferOperationResponse {}
+  interface SetOptionsOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.setOptions,
+        OperationResponseTypeI.setOptions
+      >,
+      Horizon.SetOptionsOperationResponse {}
+  interface ChangeTrustOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.changeTrust,
+        OperationResponseTypeI.changeTrust
+      >,
+      Horizon.ChangeTrustOperationResponse {}
+  interface AllowTrustOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.allowTrust,
+        OperationResponseTypeI.allowTrust
+      >,
+      Horizon.AllowTrustOperationResponse {}
+  interface AccountMergeOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.accountMerge,
+        OperationResponseTypeI.accountMerge
+      >,
+      Horizon.AccountMergeOperationResponse {}
+  interface InflationOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.inflation,
+        OperationResponseTypeI.inflation
+      >,
+      Horizon.InflationOperationResponse {}
+  interface ManageDataOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.manageData,
+        OperationResponseTypeI.manageData
+      >,
+      Horizon.ManageDataOperationResponse {}
+  interface BumpSequenceOperationRecord
+    extends BaseOperationRecord<
+        OperationResponseType.bumpSequence,
+        OperationResponseTypeI.bumpSequence
+      >,
+      Horizon.BumpSequenceOperationResponse {}
+
+  type OperationRecord =
+    | CreateAccountOperationRecord
+    | PaymentOperationRecord
+    | PathPaymentOperationRecord
+    | ManageOfferOperationRecord
+    | PassiveOfferOperationRecord
+    | SetOptionsOperationRecord
+    | ChangeTrustOperationRecord
+    | AllowTrustOperationRecord
+    | AccountMergeOperationRecord
+    | InflationOperationRecord
+    | ManageDataOperationRecord
+    | BumpSequenceOperationRecord;
+
+  interface TradeRecord extends Horizon.BaseResponse {
+    id: string;
+    paging_token: string;
+    ledger_close_time: string;
+    offer_id: string;
+    base_offer_id: string;
+    base_account: string;
+    base_amount: string;
+    base_asset_type: string;
+    base_asset_code?: string;
+    base_asset_issuer?: string;
+    counter_offer_id: string;
+    counter_account: string;
+    counter_amount: string;
+    counter_asset_type: string;
+    counter_asset_code?: string;
+    counter_asset_issuer?: string;
+    base_is_seller: boolean;
+
+    base: CallFunction<AccountRecord>;
+    counter: CallFunction<AccountRecord>;
+    operation: CallFunction<OperationRecord>;
+  }
+
+  export interface TransactionRecord
+    extends Omit<Horizon.TransactionResponse, 'ledger'> {
+    ledger_attr: Horizon.TransactionResponse['ledger'];
+
+    account: CallFunction<AccountRecord>;
+    effects: CallCollectionFunction<EffectRecord>;
+    ledger: CallFunction<LedgerRecord>;
+    operations: CallCollectionFunction<OperationRecord>;
+    precedes: CallFunction<TransactionRecord>;
+    self: CallFunction<TransactionRecord>;
+    succeeds: CallFunction<TransactionRecord>;
+  }
+
 }
