+++ types/index.d.ts
@@ -3,9 +3,8 @@
 /// <reference types="node" />
 
 import { AxiosInstance } from "axios";
-import { AssetType, Asset, Memo, MemoType, Transaction } from 'stellar-base';
+import { xdr as xdr2, AssetType, Asset, Memo, MemoType, Transaction } from 'stellar-base';
 import { Horizon } from '../src/horizon_api'
-
 // Re-StellarBase
 export {
   Account,
@@ -64,65 +63,11 @@ export namespace Config {
   function setDefault(): void;
 }
 
-interface Timebounds {
-  minTime: number;
-  maxTime: number;
-}
-
-export class Server {
-  constructor(serverURL: string, options?: Server.Options);
-  accounts(): Server.AccountCallBuilder;
-  assets(): Server.AssetsCallBuilder;
-  effects(): Server.EffectCallBuilder;
-  ledgers(): Server.LedgerCallBuilder;
-  loadAccount(accountId: string): Promise<Server.AccountResponse>;
-  offers(resource: string, ...parameters: string[]): Server.OfferCallBuilder;
-  operations(): Server.OperationCallBuilder;
-  orderbook(selling: Asset, buying: Asset): Server.OrderbookCallBuilder;
-  paths(
-    source: string,
-    destination: string,
-    destinationAsset: Asset,
-    destinationAmount: string
-  ): Server.PathCallBuilder;
-  payments(): Server.PaymentCallBuilder;
-  submitTransaction(
-    transaction: Transaction
-  ): Promise<Server.TransactionRecord>;
-  tradeAggregation(
-    base: Asset,
-    counter: Asset,
-    startTime: Date,
-    endTime: Date,
-    resolution: Date
-  ): Server.TradeAggregationCallBuilder;
-  trades(): Server.TradesCallBuilder;
-  transactions(): Server.TransactionCallBuilder;
-
-  fetchBaseFee(): Promise<number>;
-  fetchTimebounds(seconds: number): Promise<Timebounds>;
-  operationFeeStats(): Promise<Horizon.OperationFeeStatsResponse>;
-
-  serverURL: any; // TODO: require("urijs")
-}
-
 type CallBuilderResponse = Horizon.BaseResponse | Server.CollectionPage;
 
+export const HorizonAxiosClient: AxiosInstance;
+
 export namespace Server {
-  abstract class CallBuilder<
-    FetchResponse extends CallBuilderResponse,
-    StreamResponse extends CallBuilderResponse = FetchResponse
-  > {
-    constructor(serverUrl: string);
-    call(): Promise<FetchResponse>;
-    cursor(cursor: string): this;
-    limit(limit: number | string): this;
-    order(direction: 'asc' | 'desc'): this;
-    stream(options?: {
-      onmessage?: (record: StreamResponse) => void;
-      onerror?: (error: Error) => void;
-    }): () => void;
-  }
 
   interface CollectionPage<
     T extends Horizon.BaseResponse = Horizon.BaseResponse
@@ -130,60 +75,6 @@ export namespace Server {
     records: T[];
     next: () => Promise<CollectionPage<T>>;
     prev: () => Promise<CollectionPage<T>>;
-  }
-
-  /* Due to a bug with the recursive function requests */
-  interface CollectionRecord<
-    T extends Horizon.BaseResponse = Horizon.BaseResponse
-  > {
-    _links: {
-      next: Horizon.ResponseLink;
-      prev: Horizon.ResponseLink;
-      self: Horizon.ResponseLink;
-    };
-    _embedded: {
-      records: T[];
-    };
-  }
-
-  interface CallFunctionTemplateOptions {
-    cursor?: string | number;
-    limit?: number;
-    order?: 'asc' | 'desc';
-  }
-
-  type CallFunction<
-    T extends Horizon.BaseResponse = Horizon.BaseResponse
-  > = () => Promise<T>;
-  type CallCollectionFunction<
-    T extends Horizon.BaseResponse = Horizon.BaseResponse
-  > = (options?: CallFunctionTemplateOptions) => Promise<CollectionRecord<T>>;
-
-  interface AccountRecord extends Horizon.BaseResponse {
-    id: string;
-    paging_token: string;
-    account_id: string;
-    sequence: number;
-    subentry_count: number;
-    inflation_destination?: string;
-    last_modified_ledger: number;
-    thresholds: Horizon.AccountThresholds;
-    flags: Horizon.Flags;
-    balances: Horizon.BalanceLine[];
-    signers: Array<{
-      key: string;
-      weight: number;
-      type: string;
-    }>;
-    data: (options: { value: string }) => Promise<{ value: string }>;
-    data_attr: {
-      [key: string]: string;
-    };
-    effects: CallCollectionFunction<EffectRecord>;
-    offers: CallCollectionFunction<OfferRecord>;
-    operations: CallCollectionFunction<OperationRecord>;
-    payments: CallCollectionFunction<PaymentOperationRecord>;
-    trades: CallCollectionFunction<TradeRecord>;
   }
 
   interface AssetRecord extends Horizon.BaseResponse {
@@ -195,166 +86,6 @@ export namespace Server {
     num_accounts: number;
     flags: Horizon.Flags;
   }
-
-  interface EffectRecord extends Horizon.BaseResponse {
-    account: string;
-    paging_token: string;
-    starting_balance: string;
-    type_i: string;
-    type: string;
-    amount?: any;
-
-    operation?: CallFunction<OperationRecord>;
-    precedes?: CallFunction<EffectRecord>;
-    succeeds?: CallFunction<EffectRecord>;
-  }
-
-  interface LedgerRecord extends Horizon.BaseResponse {
-    id: string;
-    paging_token: string;
-    hash: string;
-    prev_hash: string;
-    sequence: number;
-    transaction_count: number;
-    operation_count: number;
-    closed_at: string;
-    total_coins: string;
-    fee_pool: string;
-    base_fee: number;
-    base_reserve: string;
-    max_tx_set_size: number;
-    protocol_version: number;
-    header_xdr: string;
-    base_fee_in_stroops: number;
-    base_reserve_in_stroops: number;
-
-    effects: CallCollectionFunction<EffectRecord>;
-    operations: CallCollectionFunction<OperationRecord>;
-    self: CallFunction<LedgerRecord>;
-    transactions: CallCollectionFunction<TransactionRecord>;
-  }
-
-  interface OfferAsset {
-    asset_type: AssetType;
-    asset_code?: string;
-    asset_issuer?: string;
-  }
-
-  interface OfferRecord extends Horizon.BaseResponse {
-    id: string;
-    paging_token: string;
-    seller: string;
-    selling: OfferAsset;
-    buying: OfferAsset;
-    amount: string;
-    price_r: Horizon.PriceRShorthand;
-    price: string;
-    last_modified_ledger: number;
-    last_modified_time: string;
-  }
-
-  import OperationResponseType = Horizon.OperationResponseType;
-  import OperationResponseTypeI = Horizon.OperationResponseTypeI;
-  interface BaseOperationRecord<
-    T extends OperationResponseType = OperationResponseType,
-    TI extends OperationResponseTypeI = OperationResponseTypeI
-  > extends Horizon.BaseOperationResponse<T, TI> {
-    self: CallFunction<OperationRecord>;
-    succeeds: CallFunction<OperationRecord>;
-    precedes: CallFunction<OperationRecord>;
-    effects: CallCollectionFunction<EffectRecord>;
-    transaction: CallFunction<TransactionRecord>;
-  }
-
-  interface CreateAccountOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.createAccount,
-        OperationResponseTypeI.createAccount
-      >,
-      Horizon.CreateAccountOperationResponse {}
-  interface PaymentOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.payment,
-        OperationResponseTypeI.payment
-      >,
-      Horizon.PaymentOperationResponse {
-    sender: CallFunction<AccountRecord>;
-    receiver: CallFunction<AccountRecord>;
-  }
-  interface PathPaymentOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.pathPayment,
-        OperationResponseTypeI.pathPayment
-      >,
-      Horizon.PathPaymentOperationResponse {}
-  interface ManageOfferOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.manageOffer,
-        OperationResponseTypeI.manageOffer
-      >,
-      Horizon.ManageOfferOperationResponse {}
-  interface PassiveOfferOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.createPassiveOffer,
-        OperationResponseTypeI.createPassiveOffer
-      >,
-      Horizon.PassiveOfferOperationResponse {}
-  interface SetOptionsOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.setOptions,
-        OperationResponseTypeI.setOptions
-      >,
-      Horizon.SetOptionsOperationResponse {}
-  interface ChangeTrustOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.changeTrust,
-        OperationResponseTypeI.changeTrust
-      >,
-      Horizon.ChangeTrustOperationResponse {}
-  interface AllowTrustOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.allowTrust,
-        OperationResponseTypeI.allowTrust
-      >,
-      Horizon.AllowTrustOperationResponse {}
-  interface AccountMergeOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.accountMerge,
-        OperationResponseTypeI.accountMerge
-      >,
-      Horizon.AccountMergeOperationResponse {}
-  interface InflationOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.inflation,
-        OperationResponseTypeI.inflation
-      >,
-      Horizon.InflationOperationResponse {}
-  interface ManageDataOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.manageData,
-        OperationResponseTypeI.manageData
-      >,
-      Horizon.ManageDataOperationResponse {}
-  interface BumpSequenceOperationRecord
-    extends BaseOperationRecord<
-        OperationResponseType.bumpSequence,
-        OperationResponseTypeI.bumpSequence
-      >,
-      Horizon.BumpSequenceOperationResponse {}
-
-  type OperationRecord =
-    | CreateAccountOperationRecord
-    | PaymentOperationRecord
-    | PathPaymentOperationRecord
-    | ManageOfferOperationRecord
-    | PassiveOfferOperationRecord
-    | SetOptionsOperationRecord
-    | ChangeTrustOperationRecord
-    | AllowTrustOperationRecord
-    | AccountMergeOperationRecord
-    | InflationOperationRecord
-    | ManageDataOperationRecord
-    | BumpSequenceOperationRecord;
 
   interface OrderbookRecord extends Horizon.BaseResponse {
     bids: Array<{ price_r: {}; price: number; amount: string }>;
@@ -379,30 +110,6 @@ export namespace Server {
     destination_asset_issuer: string;
   }
 
-  interface TradeRecord extends Horizon.BaseResponse {
-    id: string;
-    paging_token: string;
-    ledger_close_time: string;
-    offer_id: string;
-    base_offer_id: string;
-    base_account: string;
-    base_amount: string;
-    base_asset_type: string;
-    base_asset_code?: string;
-    base_asset_issuer?: string;
-    counter_offer_id: string;
-    counter_account: string;
-    counter_amount: string;
-    counter_asset_type: string;
-    counter_asset_code?: string;
-    counter_asset_issuer?: string;
-    base_is_seller: boolean;
-
-    base: CallFunction<AccountRecord>;
-    counter: CallFunction<AccountRecord>;
-    operation: CallFunction<OperationRecord>;
-  }
-
   interface TradeAggregationRecord extends Horizon.BaseResponse {
     timestamp: string;
     trade_count: number;
@@ -415,131 +122,4 @@ export namespace Server {
     close: string;
   }
 
-  interface TransactionRecord
-    extends Omit<Horizon.TransactionResponse, 'ledger'> {
-    ledger_attr: Horizon.TransactionResponse['ledger'];
-
-    account: CallFunction<AccountRecord>;
-    effects: CallCollectionFunction<EffectRecord>;
-    ledger: CallFunction<LedgerRecord>;
-    operations: CallCollectionFunction<OperationRecord>;
-    precedes: CallFunction<TransactionRecord>;
-    self: CallFunction<TransactionRecord>;
-    succeeds: CallFunction<TransactionRecord>;
-  }
-
-  abstract class AccountCallBuilder extends CallBuilder<AccountRecord> {
-    accountId(id: string): this;
-  }
-  class AccountResponse implements AccountRecord {
-    _links: { [key in 'self']: Horizon.ResponseLink };
-    id: string;
-    paging_token: string;
-    account_id: string;
-    sequence: number;
-    subentry_count: number;
-    thresholds: Horizon.AccountThresholds;
-    flags: Horizon.Flags;
-    balances: Horizon.BalanceLine[];
-    signers: Horizon.AccountSigner[];
-    data: (options: { value: string }) => Promise<{ value: string }>;
-    data_attr: {
-      [key: string]: string;
-    };
-    inflation_destination?: any;
-    last_modified_ledger: number;
-
-    effects: CallCollectionFunction<EffectRecord>;
-    offers: CallCollectionFunction<OfferRecord>;
-    operations: CallCollectionFunction<OperationRecord>;
-    payments: CallCollectionFunction<PaymentOperationRecord>;
-    trades: CallCollectionFunction<TradeRecord>;
-    constructor(response: AccountRecord);
-    accountId(): string;
-    sequenceNumber(): string;
-    incrementSequenceNumber(): void;
-  }
-
-  abstract class AssetsCallBuilder extends CallBuilder<
-    CollectionPage<AssetRecord>,
-    AssetRecord
-  > {
-    forCode(value: string): this;
-    forIssuer(value: string): this;
-  }
-
-  abstract class EffectCallBuilder extends CallBuilder<
-    CollectionPage<EffectRecord>,
-    EffectRecord
-  > {
-    forAccount(accountId: string): this;
-    forLedger(sequence: string | number): this;
-    forOperation(operationId: number): this;
-    forTransaction(transactionId: string): this;
-  }
-
-  abstract class LedgerCallBuilder extends CallBuilder<
-    CollectionPage<LedgerRecord>,
-    LedgerRecord
-  > {
-    ledger(sequence: string | number): this;
-  }
-
-  abstract class OfferCallBuilder extends CallBuilder<
-    CollectionPage<OfferRecord>,
-    OfferRecord
-  > {}
-
-  abstract class OperationCallBuilder extends CallBuilder<
-    CollectionPage<OperationRecord>,
-    OperationRecord
-  > {
-    forAccount(accountId: string): this;
-    forLedger(sequence: string | number): this;
-    forTransaction(transactionId: string): this;
-    includeFailed(value: boolean): this;
-    operation(operationId: number): this;
-  }
-  abstract class OrderbookCallBuilder extends CallBuilder<OrderbookRecord> {}
-  abstract class PathCallBuilder extends CallBuilder<
-    CollectionPage<PaymentPathRecord>,
-    PaymentPathRecord
-  > {}
-  abstract class PaymentCallBuilder extends CallBuilder<
-    CollectionPage<PaymentOperationRecord>,
-    PaymentOperationRecord
-  > {
-    forAccount(accountId: string): this;
-    forLedger(sequence: string | number): this;
-    forTransaction(transactionId: string): this;
-  }
-
-  interface Options {
-    allowHttp: boolean;
-  }
-
-  abstract class TradeAggregationCallBuilder extends CallBuilder<
-    CollectionPage<TradeAggregationRecord>,
-    TradeAggregationRecord
-  > {}
-  abstract class TradesCallBuilder extends CallBuilder<
-    CollectionPage<TradeRecord>,
-    TradeRecord
-  > {
-    forAssetPair(base: Asset, counter: Asset): this;
-    forOffer(offerId: string): this;
-    forAccount(accountId: string): this;
-  }
-
-  abstract class TransactionCallBuilder extends CallBuilder<
-    CollectionPage<TransactionRecord>,
-    TransactionRecord
-  > {
-    transaction(transactionId: string): this;
-    forAccount(accountId: string): this;
-    forLedger(sequence: string | number): this;
-    includeFailed(value: boolean): this;
-  }
 }
-
-export const HorizonAxiosClient: AxiosInstance;
